Revert "murmur_hash for state" revert to previous state implementation
This reverts commit 6bb71bef4cc43dfb803a74c0035a2a13fcf28748.

+++ moses/FF/SelPrefFeature.h
@@ -149,7 +149,7 @@ protected:
 
 class SelPrefState : public TreeState{
 public:
-	SelPrefState(TreePointer tree, uint64_t hash)
+	SelPrefState(TreePointer tree, size_t hash)
 		: TreeState(tree)
 		, m_depRelHash(hash)
 	{}
